Backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
EbookLib==0.18
lxml==5.3.0
pillow==11.0.0
PyJWT==2.9.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
